gnl workin except for last lign and no check for leaks

+++ srcs/ft_treatment_c_s_p_d.c
@@ -48,7 +48,7 @@ char    *ft_treatment_p(void *ptr)
     }
     while (i > 0)
     {
-        nbr_final[j] = base[base_conv[i - 1]];
+        nbr_final[j] = base[(int)base_conv[i - 1]];
         j++;
         i--;
     }
